Fix dependency resolution at configuration time
Iterating over the files of a `Configuration` too early is a common mistake
even for advanced Gradle users. You should never do this at configuration time.
Instead, declare the `Configuration` as an input to your task and do whatever
transformation you need when the task is executed. This can either be in a
`doFirst {}` block or by leveraging the fact that most Gradle APIs accept
Callables/Closures as input arguments.