Move test fixtures to dedicated source set
Always try to depend on as little as possible. In this example
the my-app project only needed the test fixtures of the my-lib project,
but was instead depending on all tests. This meant that a change to a
unit test in my-lib would lead to re-running all tests in my-app too.

Instead we can separate this concern into its own sourceSet and let my-app
only depend on the test fixtures. Changes to the my-lib tests no longer have
any effect on my-app.